update (#9)
* Feature/tensor refactor (#842)

* Refactoring of tensor

* Iterator bugfix

* fixed ArgMax implementation and tests + implemented for off-axis

* bugfixes to square implementation and to crossentropyloss test

* reimplemented slice as interface to by calling GetRange

* implemented slice by just using it as an interface to GetRange + fixed distances tests

* tensor_operations.cpp test now passing; except those that have been commented out for now

* minor change

* Major reduction

* Minor changes

* WiP

* argmax fixed; ml library now does not compile

* WIP

* WIP ml bugfixing ongoing

* bugfixes to argmax usage in cross entropy test

* WIP - training test not compiling

* Updating copy

* all tests pass

* merged in upstream. TSNE test deactivated until broadcast correctly implemented

* all tests pass. broadcasting is expensive because it makes local copies of tensors to avoid the const->non-const iterator issue

* tidying

* applied style

* Fixing issue with deleted `docker-images` submodule

* fixed licences

* fixed implicit return cast

* removed todo list

* removed file not meant for this commit

* implemented more iterator based maths

* updated knn for new slices

* style

* const fix

* const fix

* cutting down unnecessary forward declarations

* bugfixing slice/unsqueeze

* style

* bugfix to unitialised val & SGNS data loading

* datalodaer bugfixes

* style

* bugfixes to trig

* style

* removed tensor operations

* bugfixed flatten

* partial re-enable tensor indexing tests

* PR review fixes

* PR review fixes

* bugfix changes to cross_entropy

* fixed leftover cmakelists test

* fixed leftover cmakelists test

* casting issue

* casting issue

* bugfix tensor constructor test issue

* bugfix py_fixed_point casting

* PR cleanup

* revert submodules

* style

* bugfix vm abs function

* Minor tweak to the VM language script (#863)

* Issue #846: Key<...> (storage): Tivial fix for Wrong calc. from bits to last block & bit indexes (#865)

* update tensor benchmark slightly (#876)

* update tensor benchmark slightly

* style

* PRelu implementation with tests (#862)

* PRelu implementation with tests

* Bugfix in tensor::FIll (#881)

* Catch exception by const ref (#882)

+++ libs/core/include/core/serializers/fixed_point.hpp
@@ -22,13 +22,13 @@
 namespace fetch {
 namespace serializers {
 
-template <typename T, std::size_t I, std::size_t F>
+template <typename T, std::uint16_t I, std::uint16_t F>
 inline void Serialize(T &serializer, fixed_point::FixedPoint<I, F> const &n)
 {
   serializer << n.Data();
 }
 
-template <typename T, std::size_t I, std::size_t F>
+template <typename T, std::uint16_t I, std::uint16_t F>
 inline void Deserialize(T &serializer, fixed_point::FixedPoint<I, F> &n)
 {
   typename fixed_point::FixedPoint<I, F>::Type data;
